Improve API error handling and error message.

+++ frontend/modelarchive/src/main/java/com/amazonaws/ml/mms/archive/ModelArchive.java
@@ -26,6 +26,8 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.SocketTimeoutException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.security.DigestInputStream;
@@ -48,7 +50,7 @@ public class ModelArchive {
     public static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
 
     private static final Pattern URL_PATTERN =
-            Pattern.compile("http(s)://.*", Pattern.CASE_INSENSITIVE);
+            Pattern.compile("http(s)?://.*", Pattern.CASE_INSENSITIVE);
 
     private static final String MANIFEST_FILE = "MANIFEST.json";
 
@@ -72,16 +74,16 @@ public class ModelArchive {
         }
 
         if (url.contains("..")) {
-            throw new InvalidModelException("Relative path is not allowed in url: " + url);
+            throw new IllegalArgumentException("Relative path is not allowed in url: " + url);
         }
 
         if (modelStore == null) {
-            throw new InvalidModelException("Model store has not been configured.");
+            throw new IllegalArgumentException("Model store has not been configured.");
         }
 
         File modelLocation = new File(modelStore, url);
         if (!modelLocation.exists()) {
-            throw new InvalidModelException("Model not found: " + modelLocation.getAbsolutePath());
+            throw new IllegalArgumentException("Model not found in model store: " + url);
         }
         if (url.endsWith(".model") || url.endsWith(".mar")) {
             try (InputStream is = new FileInputStream(modelLocation)) {
@@ -138,33 +140,52 @@ public class ModelArchive {
         }
     }
 
-    private static File download(String path) throws InvalidModelException, IOException {
-        URL url = new URL(path);
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        if (conn.getResponseCode() != HttpURLConnection.HTTP_OK) {
-            throw new InvalidModelException(
-                    "Failed download model from: " + path + ", code: " + conn.getResponseCode());
+    private static File download(String path) throws IOException {
+        HttpURLConnection conn;
+        try {
+            URL url = new URL(path);
+            conn = (HttpURLConnection) url.openConnection();
+            if (conn.getResponseCode() != HttpURLConnection.HTTP_OK) {
+                throw new IllegalArgumentException(
+                        "Failed to download model from: "
+                                + path
+                                + ", code: "
+                                + conn.getResponseCode());
+            }
+        } catch (IllegalArgumentException e) { // NOPMD
+            throw e;
+        } catch (MalformedURLException | RuntimeException e) {
+            // URLConnection may throw raw RuntimeException if port is out of range.
+            throw new IllegalArgumentException("Invalid model url: " + path, e);
+        } catch (IOException e) {
+            logger.trace("", e);
+            throw new IllegalArgumentException("Failed to download model from: " + path, e);
         }
 
-        String eTag = conn.getHeaderField("ETag");
-        File tmpDir = new File(System.getProperty("java.io.tmpdir"));
-        File modelDir = new File(tmpDir, "models");
-        FileUtils.forceMkdir(modelDir);
-        if (eTag != null) {
-            if (eTag.startsWith("\"") && eTag.endsWith("\"") && eTag.length() > 2) {
-                eTag = eTag.substring(1, eTag.length() - 1);
+        try {
+            String eTag = conn.getHeaderField("ETag");
+            File tmpDir = new File(System.getProperty("java.io.tmpdir"));
+            File modelDir = new File(tmpDir, "models");
+            FileUtils.forceMkdir(modelDir);
+            if (eTag != null) {
+                if (eTag.startsWith("\"") && eTag.endsWith("\"") && eTag.length() > 2) {
+                    eTag = eTag.substring(1, eTag.length() - 1);
+                }
+                File dir = new File(modelDir, eTag);
+                if (dir.exists()) {
+                    logger.info("model folder already exists: {}", eTag);
+                    return dir;
+                }
             }
-            File dir = new File(modelDir, eTag);
-            if (dir.exists()) {
-                logger.info("model folder already exists: {}", eTag);
-                return dir;
-            }
+            return unzip(conn.getInputStream(), eTag);
+        } catch (SocketTimeoutException e) {
+            throw new IllegalArgumentException("Download model timeout: " + path, e);
         }
-        return unzip(conn.getInputStream(), eTag);
     }
 
     private static ModelArchive load(String url, File dir, boolean extracted)
             throws InvalidModelException, IOException {
+        boolean failed = true;
         try {
             File manifestFile = new File(dir, "MAR-INF/" + MANIFEST_FILE);
             Manifest manifest;
@@ -195,12 +216,12 @@ public class ModelArchive {
                 }
             }
 
+            failed = false;
             return new ModelArchive(manifest, url, dir, extracted);
-        } catch (Exception e) {
-            if (extracted) {
+        } finally {
+            if (extracted && failed) {
                 FileUtils.deleteQuietly(dir);
             }
-            throw e;
         }
     }
